Update main.py
A slight design change in the original plan related to the "Report" call. 
When the user selects Report the app will go to another frame, and from there they will select the query they want. 
Once they select the query, a new window will pop up and in that instance the appropriate frame will be called.